refactor: Name 도메인을 common 패키지로 이동
