+++ back/src/main/java/com/baba/back/baby/domain/Baby.java
@@ -1,6 +1,6 @@
 package com.baba.back.baby.domain;
 
-import com.baba.back.oauth.domain.member.Name;
+import com.baba.back.common.domain.Name;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
