+++ back/src/main/java/com/baba/back/relation/domain/RelationGroup.java
@@ -2,7 +2,7 @@ package com.baba.back.relation.domain;
 
 import com.baba.back.baby.domain.Baby;
 import com.baba.back.oauth.domain.member.Color;
-import com.baba.back.oauth.domain.member.Name;
+import com.baba.back.common.domain.Name;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
